+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - Filter `bs4g_genesis_do_nav_content` to modify the `genesis_do_nav` content between the `<nav>` tags.
+- Add Bootstrap markup search box to the navbar via the `bs4g_genesis_do_nav_content` filter.
 ### Updated
 - Update npm dependencies for security patches to build tools